requirements/base.txt
Django==4.2.7
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-ratelimit==4.1.0
whitenoise==6.6.0
gunicorn==21.2.0